local_requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
blosc2==2.0.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
chex==0.1.6
click==8.1.3
cloudpickle==2.2.1
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
distlib==0.3.6
distrax==0.1.3
dm-haiku==0.0.9
dm-pix==0.4.0
dm-tree==0.1.8
etils==1.0.0
exceptiongroup==1.1.0
execnet==1.9.0
executing==1.2.0
filelock==3.9.0
flatbuffers==2.0.7
flax==0.6.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.5.3
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.1
huggingface-hub==0.16.4
idna==3.4
immutabledict==2.2.3
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==8.10.0
jax==0.3.25
jaxlib==0.3.25+cuda11.cudnn805
jedi==0.18.2
jmp==0.0.4
jsonschema==4.17.3
kfac-jax==0.0.3
kiwisolver==1.4.4
line-profiler==4.0.3
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.4
numexpr==2.8.4
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.4
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
prompt-toolkit==3.0.37
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
pytest-xdist==3.2.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
ray==2.2.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
tables==3.8.0
tabulate==0.9.0
tbparse==0.0.7
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-probability==0.19.0
tensorstore==0.1.40
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.19.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.14.0
